fix: View of Eventos
