+++ app/src/main/res/layout/item_event.xml
@@ -1,89 +1,116 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                                                   android:layout_width="match_parent"
-                                                   android:layout_height="match_parent"
-                                                   xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <androidx.cardview.widget.CardView
-        android:id="@+id/cardView"
-        android:layout_width="0dp"
+<androidx.cardview.widget.CardView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:layout_margin="8dp"
+    app:cardCornerRadius="8dp"
+    app:cardElevation="4dp">
+
+    <LinearLayout
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="16dp"
-        android:layout_marginEnd="16dp"
-        app:cardCornerRadius="10dp"
-        app:cardElevation="10dp"
-        app:cardUseCompatPadding="true"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent">
-        <!--Contenido-->
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:padding="10dp">
+        android:orientation="vertical"
+        android:padding="16dp">
 
-            <!--Titulo-->
-            <ImageView
+        <ImageView
                 android:id="@+id/imgCertificacion"
                 android:layout_width="match_parent"
-                android:layout_height="200dp"
-                android:layout_marginTop="10dp"
-                android:layout_marginBottom="10dp"
+                android:layout_height="150dp"
+                android:scaleType="centerCrop"
                 android:src="@drawable/eventos_test" />
 
-            <!--Descripcion-->
+
             <TextView
                 android:id="@+id/name"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="10dp"
-                android:layout_marginBottom="10dp"
-                android:text="Evento"
-                android:textSize="20sp"
+                android:layout_marginTop="8dp"
+                android:text="Titulo Evento"
+                android:textSize="16sp"
+                android:textColor="#000"
                 android:textStyle="bold" />
 
             <TextView
                 android:id="@+id/description"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="10dp"
-                android:layout_marginBottom="10dp"
-                android:text="Descripcion..."
-                android:textSize="15sp" />
+                android:layout_marginTop="4dp"
+                android:text="Descripcion evento"
+                android:textColor="#666"
+                android:textSize="14sp" />
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="10dp"
+            android:orientation="horizontal"
+            android:gravity="center_vertical">
+
+            <TextView
+                android:layout_width="90dp"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="Fecha"
+                android:textColor="#666"
+                android:textSize="14sp" />
 
             <TextView
                 android:id="@+id/date"
-                android:layout_width="match_parent"
+                android:layout_width="0dp"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="10dp"
-                android:layout_marginBottom="10dp"
-                android:text="Fecha"
-                android:textSize="15sp" />
+                android:layout_weight="3"
+                android:gravity="end"
+                android:textSize="14sp"
+                android:textColor="#000"
+                android:text="dd/MM/yyyy - dd/MM/yyyy"/>
+
+        </LinearLayout>
+        <View
+            android:layout_width="match_parent"
+            android:layout_height="1dp"
+            android:background="#D3D3D3"
+            android:layout_marginTop="8dp"/>
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="4dp"
+            android:orientation="horizontal"
+            android:gravity="center_vertical">
+
+            <TextView
+                android:layout_width="0dp"
+                android:layout_marginTop="3dp"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:textSize="14sp"
+                android:textColor="#666"
+                android:text="Ubicación"/>
 
             <TextView
                 android:id="@+id/location"
-                android:layout_width="match_parent"
+                android:layout_width="0dp"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="10dp"
-                android:layout_marginBottom="10dp"
-                android:text="Ubicación"
-                android:textSize="15sp" />
+                android:layout_weight="2"
+                android:gravity="end"
+                android:textSize="14sp"
+                android:textColor="#000"
+                android:text="Queretaro"/>
 
-            <!--button for view more-->
+        </LinearLayout>
+
+
             <Button
                 android:id="@+id/btnVerMas"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="10dp"
-                android:layout_marginBottom="10dp"
-                android:background="@color/nav_bar"
-                android:text="Ver mas"
-                android:textColor="@color/white"
+                android:layout_marginTop="15dp"
+                android:backgroundTint="#CF2A2A"
+                android:textColor="#FFFFFF"
+                android:text="Ver más"
                 android:textSize="15sp" />
         </LinearLayout>
     </androidx.cardview.widget.CardView>
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
